Refactor store - file was becoming oversized
