+++ web/shapeworks/src/components/ShapeViewer/methods.js
@@ -22,6 +22,7 @@ import {
     analysisFileShown, currentAnalysisFileParticles, meanAnalysisFileParticles,
     showDifferenceFromMeanMode, cachedParticleComparisonVectors, landmarkColorList
 } from '@/store';
+import PointDrag from './dragPoints';
 
 
 const SPHERE_RESOLUTION = 32;
@@ -424,6 +425,9 @@ export default {
     },
     renderGrid() {
         this.prepareLabelCanvas();
+        if (this.currentTab === 'info' && layersShown.value.includes('Landmarks')) {
+            this.pointDrag = new PointDrag()
+        }
 
         let positionDelta = undefined;
         let viewUpDelta = undefined;
