+++ web/shapeworks/src/components/SubsetSelection.vue
@@ -2,7 +2,7 @@
 import { DataObject, Dataset } from '@/types'
 import { defineComponent, ref } from '@vue/composition-api'
 import DataList from './DataList.vue'
-import { selectedDataObjects, allDatasets } from '../store';
+import { selectedDataObjects, allDatasets } from '@/store';
 import { subsetDataset } from '@/api/rest';
